Add variant prop to avatar
Replaces dark/light boolean props
