+++ packages/docs-site/src/library/pages/components/avatar.md
@@ -5,7 +5,7 @@ header: true
 draft: false
 ---
 
-import { Avatar, Tab, TabSet } from '@royalnavy/react-component-library'
+import { Avatar, AVATAR_VARIANT, Tab, TabSet } from '@royalnavy/react-component-library'
 import DataTable from '../../../components/presenters/data-table'
 import CodeHighlighter from '../../../components/presenters/code-highlighter'
 import SketchWidget from '../../../components/presenters/sketch-widget'
@@ -51,12 +51,11 @@ The Avatar component is a simple component that displays the users initials with
 
 ### Basic Usage
 
-<CodeHighlighter source={`
-<Avatar light initials="RT"/>
-<Avatar dark initials="ST"/>
+<CodeHighlighter source={`<Avatar initials="RT" variant={AVATAR_VARIANT.DARK} />
+<Avatar initials="ST" variant={AVATAR_VARIANT.LIGHT} />
 `} language="javascript">
   <div style={{ background: '#A0A0A0', padding: 20 }}>
-    <p><Avatar light initials="RT"/>&nbsp;<Avatar dark initials="ST"/></p>
+    <p><Avatar initials="RT" variant={AVATAR_VARIANT.DARK} />&nbsp;<Avatar initials="ST" variant={AVATAR_VARIANT.LIGHT}/></p>
   </div>
 </CodeHighlighter>
 
@@ -71,13 +70,6 @@ The Avatar component is a simple component that displays the users initials with
     Description: 'Custom css class to add to the Avatar element',
   },
   {
-    Name: 'dark',
-    Type: 'boolean',
-    Required: 'False',
-    Default: '',
-    Description: 'Use the dark style of avatar',
-  },
-  {
     Name: 'initials',
     Type: 'string',
     Required: 'True',
@@ -85,11 +77,11 @@ The Avatar component is a simple component that displays the users initials with
     Description: 'The users initials, a maximum of 2 letters',
   },
   {
-    Name: 'light',
-    Type: 'boolean',
+    Name: 'variant',
+    Type: 'string',
     Required: 'False',
     Default: '',
-    Description: 'Use the light style of avatar',
+    Description: 'Variant of the avatar, can be `dark` or `light`.',
   },
 ]} />
 </Tab>
