+++ packages/react-component-library/src/components/Avatar/Avatar.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { Avatar } from './index'
+import { Avatar, AVATAR_VARIANT } from '.'
 
 const stories = storiesOf('Avatar', module)
 
@@ -12,11 +12,11 @@ stories.add('Default', () => (
 ))
 stories.add('Light', () => (
   <div style={{ background: '#A0A0A0', padding: 20 }}>
-    <Avatar light initials="AT" />
+    <Avatar initials="AT" variant={AVATAR_VARIANT.LIGHT} />
   </div>
 ))
 stories.add('Dark', () => (
   <div style={{ background: '#A0A0A0', padding: 20 }}>
-    <Avatar dark initials="AT" />
+    <Avatar initials="AT" variant={AVATAR_VARIANT.DARK} />
   </div>
 ))
