+++ packages/react-component-library/src/components/Avatar/constants.ts
@@ -1,0 +1,6 @@
+enum AVATAR_VARIANT {
+  DARK = 'dark',
+  LIGHT = 'light',
+}
+
+export { AVATAR_VARIANT }
